Changed all models classes to return either 'OptimalSolutionFound' if all optimisation found an optimal solution or - if not the case -  the last SolverReturnStatus which failed.
This is OK to start with but I am wondering whether it would not be better to perhaps report the SolverReturnStatus per DMU in the solution. This might hit performances so perhaps instead give an ArrayList of DMU numbers for which the optimisation failed along with the optimisation SolverReturnStatus (now I am wondering whether the optimisation could fail to find an optimal solution for one DMU but could potentially find an optimal solution for other DMUs. All the time I solved DEA problems and there was some problems, ALL optimisations failed).

I think it is OK to start with this and potentially change it in future if this is not satisfactory (hence the commit).